pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-to-mattermost"
version = "1.0.0"
description = "Convert Telegram exports to Mattermost import format"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "ETH Zurich IT Security Center"},
    {name = "Chad Phillips"},
]
dependencies = [
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
telegram-to-mattermost = "telegram_to_mattermost.migrate:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.setuptools.packages.find]
where = ["."]
